src/QuickTest_Android/app/src/main/java/aplicacion/android/danielvm/quicktestandroid/models/moodle/Parameter.java
package aplicacion.android.danielvm.quicktestandroid.models.moodle;

/**
 * Clase Parameter encargada de encapsular el contenido de la respuesta JSON proporcionada
 * por el web service de Moodle en un objeto Java.
 *
 * @author Daniel Puente Gabarri.
 */

public class Parameter {

    private String name;
    private String value;

    public Parameter(String name, String value) {
        this.name = name;
        this.value = value;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getValue() {
        return value;
    }

    public void setValue(String value) {
        this.value = value;
    }
}
